tests/end_to_end/requirements.txt
pytest
dulwich
PyYAML
docker-py
